Challenge - Endpoint /order

+++ app/back-end/src/database/migrations/20220514232020-create-buyers.js
@@ -127,7 +127,7 @@ module.exports = {
       },
     
       confirm: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.INTEGER,
         defaultValue: 1
       },
